requirements_test.txt
# Strictly for tests
# https://github.com/MatthewFlamm/pytest-homeassistant-custom-component/releases
pytest-homeassistant-custom-component==0.13.85
